+++ marcade.py
@@ -14,12 +14,13 @@ Change log: Check CHANGELOG.md file.
 
 try:
     # Required modules
+    import argparse
     import os
     import sys
-    import argparse
     import pygame
     from pygame.locals import *
     # Myself modules
+    from asteroids import Asteroids
     from pong import Pong
 except ImportError as err:
     print("Could not load module. " + str(err))
@@ -32,73 +33,62 @@ class UserArgumentParser():
     http://chase-seibert.github.io/blog/
     """
     def __init__(self):
-        # Program description
-        self.program_name = "pong"
-        self.program_version = "0.2"
-        self.program_date = "2018-10-24"
-        self.program_description = "Pong"
+        self.program_name = "marcade"
+        self.program_version = "0.4"
+        self.program_date = "2018-10-27"
+        self.program_description = "MArcade"
         self.program_copyright = "Copyright (c) 2014-2018 Marcio Pessoa"
         self.program_license = "GPLv2"
         self.program_website = "http://pessoa.eti.br/"
         self.program_contact = "Marcio Pessoa <marcio.pessoa@gmail.com>"
-        header = ('run \n\n')
+        header = ('marcade <game> [<args>]\n\n' +
+                  'Games:\n' +
+                  '  asteroids      amazing Asteroids space game\n' +
+                  '  pong           classical Pong game\n\n')
         footer = (self.program_copyright + '\n' +
                   'License: ' + self.program_license + '\n' +
                   'Website: ' + self.program_website + '\n' +
                   'Contact: ' + self.program_contact + '\n')
         examples = ('examples:\n' +
-                    '  run\n')
+                    '  marcade asteroids\n' +
+                    '  marcade pong --fullscreen\n')
         self.version = (self.program_name + " " + self.program_version + " (" +
                         self.program_date + ")")
         epilog = (examples + '\n' + footer)
-        # Argument Parser
         parser = argparse.ArgumentParser(
             prog=self.program_name,
             formatter_class=argparse.RawDescriptionHelpFormatter,
             epilog=epilog,
             add_help=True,
             usage=header)
+        parser.add_argument('command', help='command to run')
         parser.add_argument('-V', '--version', action='version',
                             version=self.version,
                             help='show version information and exit')
-        # When no arguments were supplied
         if len(sys.argv) < 2:
             self.pong()
             sys.exit(False)
         args = parser.parse_args(sys.argv[1:2])
-        # When a unrecognized arguments were supplied
         if not hasattr(self, args.command):
-            echoln('Unrecognized command')
+            print('Unrecognized command')
             parser.print_help()
             sys.exit(True)
         getattr(self, args.command)()
 
-    def __start(self):
+    def __screen_start(self):
         self.window_title = self.program_description
         self.running = True
         self.screen_rate = 30  # FPS
         self.canvas_size = (480, 320)  # (width, height) pixels
         self.__screen_set()
         self.__ctrl_set()
-        self.pong = Pong(self.screen)
-        self.pong.start()
-
-    def __run(self):
-        while self.running:
-            self.__check_event()
-            self.pong.run()
-            self.clock.tick(self.screen_rate)
-            pygame.display.flip()
 
     def __screen_set(self):
         # Window position
         os.environ['SDL_VIDEO_CENTERED'] = '1'
         # Initialise screen
         pygame.init()
-        self.screen = pygame.display.set_mode(self.canvas_size,
-                                              HWSURFACE |
-                                              DOUBLEBUF |
-                                              RESIZABLE)
+        self.__screen_reset()
         # Window caption
         pygame.display.set_caption(self.window_title)
         # Clockling
@@ -109,7 +99,13 @@ class UserArgumentParser():
                                               HWSURFACE |
                                               DOUBLEBUF |
                                               RESIZABLE)
-        self.pong.size_reset()
+
+    def __run(self):
+        while self.running:
+            self.__check_event()
+            self.game.run()
+            self.clock.tick(self.screen_rate)
+            pygame.display.flip()
 
     def __check_event(self):
         for event in pygame.event.get():
@@ -121,14 +117,32 @@ class UserArgumentParser():
             elif event.type == VIDEORESIZE:
                 self.canvas_size = event.dict['size']
                 self.__screen_reset()
-            self.pong.control(event)
+                self.game.size_reset()
+            self.game.control(event)
 
     def __ctrl_set(self):
         # Set keyboard speed
         pygame.key.set_repeat(1, 1)
 
     def pong(self):
-        self.__start()
+        parser = argparse.ArgumentParser(
+            prog=self.program_name + ' pong',
+            description='classical Pong game')
+        args = parser.parse_args(sys.argv[2:])
+        self.__screen_start()
+        self.game = Pong(self.screen)
+        self.game.start()
+        self.__run()
+        sys.exit(False)
+
+    def asteroids(self):
+        parser = argparse.ArgumentParser(
+            prog=self.program_name + ' asteroids',
+            description='amazing Asteroids space game')
+        args = parser.parse_args(sys.argv[2:])
+        self.__screen_start(args.fullscreen)
+        self.game = Asteroids(self.screen)
+        self.game.start()
         self.__run()
         sys.exit(False)