+++ CHANGELOG.md
@@ -1,13 +1,21 @@
 # Change log
 All notable changes to this project will be documented in this file.
 
+## [0.4] - 2018-10-27
+### Changed
+- Project renamed from Pong to MArcade.
+
+## [0.3] - 2018-10-26
+### Added
+- Pong: Window resizing support.
+
 ## [0.2] - 2018-06-20
 ### Fixed
-- Ball spawn to never raize balls without angle.
+- Pong: Ball spawn to never raize balls without angle.
 
 ## [0.1] - 2014-06-06
 ### Added
-- First version.
+- Pong: First version.